Added a check of the output script length
This chceck verifies whether the parsing of the block hex string ends precisely at its end and if not, it marks the block as invalid and avoids running into problems (unhandled IndexOutOfRange exception in the case of the parsed length being incorrectly greater than the real length). By extension, this verifies whether the parsed length of the output script corresponds to its real length.